requirements.txt
Flask
Werkzeug
firebase-admin
openai
bcrypt
pyrebase4
gunicorn


